Refactor: 대시보드 페이지 단일화
